requirements.txt
mcp[cli]
pycryptodome
sympy
requests
httpx
bcrypt
